for bigger examples, like cyclic-10, we need 64bit hash values

+++ src/types.h
@@ -65,7 +65,7 @@ typedef uint32_t nelts_t;
 typedef uint32_t ht_size_t;
 
 /* hash table entry size */
-typedef uint32_t hash_t;
+typedef uint64_t hash_t;
 
 /* degree size */
 typedef uint16_t deg_t;
